Refactoring: GeoDataRiverSurvey::MouseRotateRiverCrosssectionCommand is moved to separated files.
